Factor out LazyWrapper & rsync classes from util.py.
- Move LazyWrapper into its own util module. This allows other modules
to use LazyWrapper without pulling all the dependencies that
gslib/util.py imports.
- Similarly, shared rsync functionality should be in its own module with
minimal dependencies.
